fix(disclosureDashboard): replace dropdown component
