+++ src/components/atoms/localisationIssues/localisationIssues.jsx
@@ -1,0 +1,8 @@
+import React from 'react'
+
+const LocalisationIssues = ({cssClass, text }) => {
+
+return <span className={`localisation-component ${cssClass}`}>{text}</span>
+}
+
+export default LocalisationIssues;
