+++ src/components/atoms/localisationIssues/localisationIssues.scss
@@ -1,4 +1,4 @@
-.localisation-incidents {
+.localisation-issues {
   @include disclosure-information;
   color: $secondary-utility-color;
 }
